Make CONSTANTLY get a source transform again.
Commit ca8135a6852cde2206ce9bdaa9b9d57f3b047f4e said, "Replace the
source-transform with a DERIVE-TYPE optimizer: there are essentially
no cases where the source transform improves performance."

However, this was wrong. The source transform was crucially important
for transforming the case of (FUNCALL (CONSTANTLY 4)), which stopped
getting optimized, causing someone to file a bug for it later, which
then caused this optimization to be reinvented with an IR1
transform. However, the source transform does the same job but is much
simpler, so restore it and get rid of the transform and derive-type,
which is made redundant by the source transform. It is also faster to
allocate and reference a component local closure than calling out to
the CONSTANTLY, so it is better to do the transform in all cases, not
just with funcalls.