Make stack allocating anonymous closures work.
* Introduce a DX-INFO structure to do the bookkeeping better. This way
we don't need to do list hackery anymore during optimization, and we
don't need to carry the DX kind on the cleanup.
* Let-bound anonymous closures get marked as being dynamic extent when
the LET variable gets substituted away with propagate-ref-dx.
* Simplify RECOGNIZE-POTENTIALLY-DYNAMIC-EXTENT-LVARS a bit so we
don't need to iterate over the DX lvars twice.
* Simplify comments in DETERMINE-LAMBDA-VAR-AND-NLX-EXTENT: It tried
to predict how this change would be implemented, but we just mark the
extent of the main entry rather than the XEP when propagating DX, so
the cases become uniform and we can simplify the logic and not rely on
clever punning.