Improve and simplify downward dxification.
Don't create wrappers and transform to DX-FLET - we have the ability
to directly mark the lvars and stack allocate anonymous closures
now. As a result we can lift the restriction on only mandatory args
for the function being DXed. By making this no longer a transform, we
also open up the possiblity of doing this marking in some other phase
of the compiler (which it probably should be) as well as maybe using
it for known functions as well.

Also ignore non-closures when noting no stack allocation.