Fix TAIL-ANNOTATE for the nth time.
There was in fact something wrong with every iteration of the comments
over the past 20 years. NS was right to want to make the XEP not
appear - XEPs are just the trampolines and showing them would make the
function appear twice. However, what actually inhibits the XEPs
appearance is always annotating tail-recursive *local calls* as tail,
even when it's non-returning, since XEPs local call the actual
function.

I think people got confused because of these two facets - there's the
XEP, as well as the main function. The XEP should always tail-call the
main function because we don't want it to show up in the backtrace
unless there's an actual error happening within the XEP (think
optional processing problems etc).

Also remove ALLOW-NON-RETURNING-TAIL-CALL. Not only is the function it
was originally written for gone, but it was inherently bogus: It
achieved function frame disappearance by making all non-returning
calls tail calls, even when they aren't in tail position. If the
intention is just to make the function not appear in the backtrace,
the debugger already has a facility to do that. Functions that just
call error do not need TCO per se, because, well, they aren't going to
consume more stack frames and don't need to be fast. Create a macro
helper to express the intent more clearly. We also avoid scanning
every single node in the component during tail annotation.

Also, some non-trivial functions were wrongly annotated with
ALLOW-NON-RETURNING-TAIL-CALL, hiding potentially useful local
variables. So, don't make those error wrappers.