Make auto downward dxification work for known functions too.
This is the real payoff of the new implementation, as known functions
tend to be the more common case of auto dxification (think MAP and
friends). Casts don't stop dx propagation so it works with known
functions as well. Many more closures get dynamic extent allocated
during self build now.

* Move the dx marking to environment analysis, so that we don't
repeatedly try to dxify the same combination, especially so that we
don't interact with unknown key args becoming known as much. This also
unifies the case for known functions, which we enable now.

* We add some more truly-dynamic-extent declarations to squash some
notes now that more dx declarations take effect for real.

* Also use a more logical criteria when checking if it actually makes
sense to dx propagate to a closure. Being in the same stack frame is
what's necessary, not some lexical property of the lambda containing
lexenv.

* Use a broader definition of what counts as "probably not going to
get redefined", so that the behavior on #+sb-devel is a bit more
similar to #-sb-devel.

Upshot is that we now dynamic extent allocate as many closures as the
original CMU CL implementation of dynamic extent did. (Which wasn't
stealable in 2003 because CMU CL doesn't have the concept of ENCLOSE,
making it easier to do the auto detection. We also have the benefit of
specifying dynamic extent per arg, rather than per known function.)

I think we are still missing some dynamic extent declarations
though. Some missing ones could probably be taken by looking for the
dynamic extent safe attribute in CMU CL's fndb.lisp.