Make auto-dx-ification with transformed known calls work.
Do this by doing cleanup insertion and marking right before we
transform, and passing the already created cleanups along when the
same closure is detected as being used as a funarg to the transformed
call (the "no other refs" condition will prevent the closure from
being used by unrelated calls.)

This is extremely rube-goldberg-machine-esque but gets the job
done. This eliminates most unactionable "can't stack allocate" notes
in self-build and adds some new ones which *are* actionable. We need
to rework how stack allocation notes work, but basically the new ones
are saying that the user is not passing a DX function to a known
function which takes one. (i.e. the compiler won't auto-stack-allocate
it because it can't prove it's safe to do so, and detects that you are
passing it without declaring it dynamic extent). The note for this
situation should really say something like "passing a heap-allocated
closure to a function that only uses it as a downward funarg."