Use the correct mess-up node when propagating dx to lambdas.
If there is an entry node, that needs to be the mess-up, so that stack
analysis doesn't get confused where the DX environment begins. (I
tried to short-cut this by making the enclose the mess-up in one of
the propagate-ref-dx cases but forgot about stack analysis
implications.) This requires adding another case in environment
analysis specifically when we don't propagate through a REF as
well. downward-dx-ification was also bashed a bit with trial and error
to find something that worked.

However, the picture is becoming clearer how to deal with all these
disparate functional dx propagation cases more uniformly, so once the
bugs are ironed out and tests are added there are probably some
redundant cases to be unified or cleaned up entirely.