Remove entry-point-label function
The idea was to convert jumps between assembly routines into the same
instruction form as if it had been to a labeled statement. We can do that
using ordinary fixups but resolving each fixup into a label in assemfile
rather than as an address in genesis. This makes other fixups such as those
inserted by ALLOCATION when calling ALLOC-TRAMP look as though you had used
a label when compiling SIGNED-BIGNUM-IN-Rxx etc, so that the assembler will
choose "call rel32" form instead of "call [ea]" form.