Close over closures more efficiently.
Do it as a separate pass after all lambda vars and NLX-INFOs are
closed over, rather than trying to have CLOSE-OVER keep track of the
environments that closing over closures touches. We cons less
intermediate data and do less membership checking, which was done to
prevent recursion for mutually referential closures.

The code for CLOSE-OVER also becomes simpler and matches the comment
for it better.