x86-64: add another way to avoid POSIX signals for internal errors
In you case you need to debug code which actually gets a SIGILL,
so UD2 or INTO aren't the best choice of trap instruction.
This doesn't pass the regression suite, but it's enough to help debug
a crashing self-build using gdb.
Add --without-int3-breakpoints in make-config to use this feature.
Works only if #+linux at present.