Reduce unusable heap space due to WITH/WITHOUT-ARENA
alloc_start_page is the maximum page index used by any number of threads
opening and closing allocation regions. So if gc_close_region is invoked
more often than it normally would be, the next find_freeish_page tends to
leave a lot of waste behind it. This is mitigated by having each thread
remember where it would like to pick up. This heuristic works extremely
well and has no down-side to it.