Never anonymize constants in IR1.
Thus always preserving the names in the front end, as intended (see
comment describing the CONSTANT structure). We do anonymize when
translating to IR2, because we don't want duplicate code constants for
anonymous and named constants with the same value in the IR2 (and
hence code) component.

Ideally we could get rid of the source name slot in the REF structure
now, but something I don't really understand in debug-dump uses it.