Remove stub for %MAKE-FUNCALLABLE-INSTANCE
Failure of the code/data-separation test under the feature
set #+(and executable-funinstances (not compact-instance-header))
revealed that the fixed-alloc vop used a wrong feature test; and that
further implied that the var-alloc vop had to be wrong since it never
had a special case for funinstances. If a funinstance were on a GC page
not flagged as PAGE_TYPE_CODE, then it could not undergo implicit pinning
by a PC, and in the worst case might crash if GC tried to free or move
the embedded trampoline while executing it.

It turns out there is no compelling need to make variable-length
funcallable instances.  At least according to the test suite.