editcore: Scan funcallable instances correctly.
The current logic made very little sense: funcallable instances do not
really support untagged slots and otherwise it is completely broken on
non-executable fun-instances. Also refactor the control flow to be
less assignment-heavy and more understandable. The compiler also is
able to assignment-convert SCAN-OBJ this way, transforming this
function into a loop.