Remove x86-64 specific optimization (non-x86-64 pessimization).
The micro-optimization it enables on x86-64 is actually a
pessimization on other platforms, where it is trivial to do memory
operations with NIL (in a register) and UNBOUND-MARKER (which is an
offset off of NULL-TN). Loading bit patterns is more expensive on
those platforms. More importantly, hardwiring bit values precludes
static space being relocatable, and also if x86-64 gets NULL-TN (a
pre-requisite to static space being relocatable) it would also be a
pessimization there.