Add relocatable-static-space feature,
Enabling this feature allows static space to be relocatable on
start-up. It currently works only on arm64 (a small bit of assembly
needed to load the value of NIL is needed), but any platform that has
NULL-TN can trivially implement this feature as well by loading NIL as
STATIC_SPACE_START + NIL_VALUE_OFFSET.

So x86-64 could have this feature as well if it had NULL-TN, but a few
very micro-optimizey tricks there would need to be removed as well.

Currently use of r/o space is disabled when static space relocation is
possible, because we would need to fix-up the pointers to static space
in r/o space on start-up, negating the point of having it be r/o. I
left a FIXME so we know to do this more correctly at some point in the
future.