Never resignal synchronous low-level signals
I can't see why it would have been expected to work on SIGFPE from a foreign
thread so I guess most C code does not use floating-point exceptions.
And in particular, given that assert() can raise SIGABRT, don't try to handle
a failed assertion from a foreign thread in a Lisp thread.

I initially thought that the strange behavior must have been a recent change
but it seems to have come from rev 9ee246f5 which originally added re-signaling,
and always considered it reasonable to do regardless of whether the signal was
process-directed or thread-directed. I think there is an indicator in siginfo_t
of which it was, but this is adequate for now.