Stub out read barrier support for weak hash-tables
In the same vein as git rev 68540021b9 for weak-pointer read barriers, this
affects weak hash-tables but without a change to the primitive object holding
the k/v pairs. The representation of a table is opaque to users so there's no
concern that someone will call SVREF wrongly. To signify where a barrier would
be emitted, this change just substitutes WEAK-KVV-REF for SVREF in the few
places that need it, which degenerates to SVREF if weak-vector-readbarrier is
not enabled. And when enabled, for now it's still SVREF but only until vops are
added for mutual exclusion with clearing of weak refs.