Move a bunch of preprocessor macros out of gc headers
The goal is to have good separation between macros that deal with object
representation versus the allocator/collector and other random logic.
Previous attempts have been a total failure, considering that there are
nearly 30 files that include gc-private.h. By starting over with the
manipulators of the object model, it should be possible to arrive at a
healthier state with respect to changing out the GC.