Remove and/or rename a lot of #+gencgc
The C runtime will soon need to support the new mark-region collector which
is generational in nature with the identical heap structure as gencgc.
For clarity's sake, :GENCGC will refer to the specific implementation
contained in the file gencgc.c, whereas :GENERATIONAL shall imply either
gencgc or :MARK-REGION-GC without having to spell the (OR ...) everywhere.
The new feature name is not perfect- there could exist an incompatible
generational GC. Deal with that later, if ever.

That said, a few techniques were used in changing featureization:
- If it was code we would like to work on any GC, then delete the conditional.
  Later, conditional code could be re-added if something can't be made to work
  for whatever different GC is implemented.
- If it was something such as GENERATION-OF we'd like to remember not to allow
  to be called with a non-generational collector, add #+generational.
- If something was #-gencgc but meant "#+cheneygc" then either change to that
  or delete the code. Retained dead code is a hint to where subsequent changes
  might be required if not obvious, for example the compiler's decision about
  L-T-V and value-cells. Absence of a feature marker could leave you wondering.
- No machine-dependent files were touched, as the mark-region collector
  will support only x86-64 at present.

(I don't always write long commit messages, but when I do, they concern
multiple work-in-progress GC implementations.)