stack: Rewrite and simplify stack liveness analysis.
This new implementation makes back propagation look more like a normal
liveness algorithm by only propagating lifetimes from "DESTs" (really
where the next enclosing allocation begins) to USEs, avoiding all the
trouble caused by conditional allocation in the old framework, since
now the graph walks are bounded by the live ranges of the LVARs
instead of the ranges bounded by the allocations to the beginning of
the lexical scope of the object. We then do an extra pass to endow the
live sets through dynamic extent semantics. This second pass doesn't
need to fixpoint because it's not dependent on predecessors or
successors and hence isn't susceptible to recomputation around
loops. This algorithm doesn't require the use of dominators, avoids
extra bits of consing, is more obviously correct, and does much less
repeated work since we don't need to do a graph walk *during* the
solving of the dataflow equations and also reduces the amount of set
merging and differencing substantially.