stack: Preserve the stack below DX objects when ordering sets.
Instead of doing it when computing liveness sets. Since the condition
for preserving stack values below DX objects on the stack relies on
the ordering of the stacks, it's less brittle to just apply this
condition when sorting the stacks, instead of trying to figure out
what is underneath what on the stack when computing liveness using the
state of the start stack and the information in the PUSHED set for
every subpart of a DX object. Also add some exposition of why this
works at all with the new state of things where we just push one lvar
on the stack per group of dynamic extent objects allocated together,
since we were already assuming this worked. I hope it's correct, and
that the ordering invariants checked for during set ordering catches
any badness combined with the checks in things like %POP- and
%NIP-VALUES as well as the conversion of mv-call in the backend.