SB-POSIX:SYSCALL-ERROR for routines that can return NULL.
Specifically: getpwnam, getpwuid, getgrnam, getgruid, readdir, and
getenv. For all but getenv, POSIX describes the convention that a NULL
return value and "unchanged" errno is a non-error (including
end-of-directory for readdir, no-such-user for getpw*, etc), but NULL
and a changed errno is an error. SB-POSIX hasn't ever signaled errors
for a changed errno before, but probably should. (As for getenv, POSIX
does not define any errors for it, but FreeBSD and OSX document
setting errno to EINVAL sometimes; and FreeBSD actually behaves as
documented; so it seems reasonable to bind to whatever the OS does.)

Unfortunately, it's unclear how to force these errors for testing. For
the ones having to do with the passwd and group databases, SBCL
probably can't and certainly shouldn't try to interfere with them. For
readdir, I was unable to induce an error in a few experiments
concurrently modifying a directory while SBCL was enumerating it; and
the getenv case isn't required to set errno anywhere.

committer: Douglas Katzman <dougk@google.com>