Add WITH-PASSWD-DATABASE and WITH-GROUP-DATABASE to SB-POSIX
These protect against concurrent uses of the respective databases in
distinct threads, as well as prohibit attempts at reentrant use. It's
up to users to wrap their getpw* and getgr* calls in these macros.

(No tests are added for these because it seemed as if doing so would
merely exercise WITH-MUTEX, which is tested for elsewhere.)

committer: Douglas Katzman <dougk@google.com>