Commit some files from SML# release 3.5
This is the basis of the work-in-progress concurrent GC for SBCL.
We don't want the latest SML# runtime because it depends on the authors'
user-mode thread library which has not been investigated for its suitability
toward providing Lisp-specific needs such as special variables,
in addition to not being ported to all architectures SBCL supports.
Therefore we'll have a better time interfacing to a slightly older
version of their GC that is every bit as effective for our purposes.

SML# is licensed under 3-clause BSD which is compatible with SBCL.
The files in this commit are unmodified from SML# git rev 681f44a816d1
other than converting tabs to spaces.
Full source available at https://github.com/smlsharp/smlsharp.git