Ensure boot-time effective methods also perform keyword checks
The effective methods generated during the PCL bootstrap are cached in
surprisingly sticky ways, not even invalidated through a
REINITIALIZE-INSTANCE of the generic function.  Adjust the secondary
dispatch function generation to wrap the keyword check if necessary
around boot-time effective methods.

This in combination with the previous change causes the valid keyword
computation to happen on early generic functions: ensure that
COMPUTE-APPLICABLE-KEYWORDS can work with those.