Don't (yet) check keywords for already-error-case effective methods
Because what our CLOS implementation needs is MORE MAGIC :-/

We currently handle the cases of no primary methods, and invalid
qualifiers in short method combinations, using specially-recognized
function-like calls to not-existing functions.  We've known that this
is wrong from the point of view of MOP users for a long time; there's
a comment in SHORT-COMPUTE-EFFECTIVE-METHOD explaining the problem.
For the purposes of this freeze period, though, preserve the magic by
not introducting keyword argument checks if we know we are already
dealing with one of these special error-operator-based effective
methods.