Use some magic from elsewhere
Although ENSURE-FNGEN can't safely construct a dumpable form of a
quoted list of methods, it can handle (by argument passing) a list of
quoted methods.  Rearrange the METHOD-COMBINATION-ERROR calls in
standard- and short-compute-effective-method to be of this form.  This
allows precom2.lisp to compile successfully, even when system method
definitions contain argument combinations where there will be no
primary method.

This also opens the possibility of inserting keyword argument checks
in the secondary dfuns for no-applicable-method / no-primary-method.