Reduce the magic in COMPUTE-EFFECTIVE-METHOD
Instead of returning implicitly special forms from
COMPUTE-EFFECTIVE-METHOD for cases of no-primary-method and
method-with-invalid-qualifiers, signal dedicated conditions to cause a
non-local-exit at the caller, in the case that we are calling from our
own machinery.  If we are not, fall back in effective method
computation to various invalid-method or method-combination errors, so
that whatever is calling us to generate the effective method can deal
with them appropriately.

This works, but is a bit confused about when things should happen.
INVALID-QUALIFIERS actually shouldn't need any magic at all, because
we have everything available to generate a valid (error-generating)
effective method.  NO-PRIMARY-METHOD needs additional support, but
that can be provided by doing a check of the applicable methods before
calling COMPUTE-EFFECTIVE-METHOD (see the next change).