Fix UNREAD-CHAR for input streams with no character buffer
We must keep track of the previous octet buffer position so that
UNREAD-CHAR can revert to that position.  (Only one previous position
is needed because UNREAD-CHAR is specified to require reads between
successive calls to UNREAD-CHAR; effectively we only need to provide
one-character lookback.)

If we read a character from INSTEAD, make sure that PREV-HEAD is set
to HEAD, and if UNREAD-CHAR is called when the octet buffer is in that
state, then the previous read must have taken from the INSTEAD vector
so the unread must push to it.