Fix ludicrous BYTE-BLT usage convention
In almost all cases the starting destination index is 0, making the ending index
equivalent to a byte count, as it should be. But not only was it unobvious why
it didn't elect to take a difference of source indices to imply a length, neither
way is particularly better. How about just don't do extra arithmetic?