Part 2 of ELF core for mark-region
Using the following recipe, you can get a core with code split out into
the contiguous text space:
* (load "tools-for-build/editcore")
* (sb-editcore::move-dynamic-code-to-text-space "output/sbcl.core" "new.core")
* (quit)
$ mv new.core output/sbcl.core

The recipe works for #+(and gencgc (not immobile-space)) also,
though a bunch of regression tests fail for the resulting core,
whereas with mark-region no test fails that didn't already.

Converting the text space into an ELF section will be the next step.