ppc: add vop for SB-LOCKLESS:GET-NEXT
The inline definition was inadequate for some out-of-order CPUs
as evidenced by frequent failure in the solist.pure tests.

On ppc64, GC calls preserve_pointer() for all context registers, and
a word ending in #b000 pointing to INSTANCE_WIDETAG acts as if it
had INSTANCE_POINTER_LOWTAG. So this change is not needed there
because untagged NODE-NEXT pointers just work with no fuss.