Use symbol-hash directly in more places
Firstly there is no need to check if the hash slot was set, so the transform
that used to insert an IF around ensure-symbol-hash is irrelevant.
Secondly, after implementing the work-in-progress change alluded to in the
comment above SYMBOL-TABLE-HASH which asks for symbols to get some random hash
bits as well as a name-based hash, then anywhere we want to use all hash bits
needs SYMBOL-HASH. Continue to use SXHASH to retain the previous behavior that
STRING= names imply = hashes, and that exact numerical value are fixed across
separate runs of SBCL as CLHS seems to require.