Improve the structure-typecase expansion
Use the perfect hash generator rather than rely on an expansion which can
resolve collisions after masking layout-clos-hash to a reduced width.
The expander is nearly agnostic of whether it's cross-compiling - there is
a record/playback of the hash functions made during a native build.

This change compiled fine using CCL and ECL and older SBCL as hosts.