Improve FIND,POSITION transform on constant sequences
Do not expand into CASE. CASE had to do a lot of exceess work to discern
that this is an operation that can not invoke arbitrary user code.
There is literally no control flow, apart from an array bound check.
Moreover, the old way required a single range of SYMBOL-HASH bits to choose
the outcome (the constraint being that PICK-BEST-SXHASH-BITS must return 1,
for a single LDB step), but the new transform uses a perfect hash.
If we could compile that to as few instrutions as claimed in comments
throughout perfecthex.inc it would be even better.