Convert a FLET outside an expansion to a MACROLET inside
This is motivated by attempting to make slightly more reproducible
fasls under CLISP, though the exact mechanism for why the MACROLET is
reproducible while the FLET is not is unclear.

The immediate symptom of the unportability is that the
COMPILED-DEBUG-FUN structure dumped for (FLET @ :IN BIGNUM-REPLACE)
has two gensymed VARs with names G<nnnnn> for different nnnnn
depending on the build host.

It was not at all clear to me after an hour or so of staring at this
either where these VARs were coming from, or why they differed in the
SB-XC:*GENSYM-COUNTER* at the time of their creation.  (I did manage
to verify that they were generated by SB-XC:GENSYM and not
CL:GENSYM).  However, we can get around the symptom by generating the
code with a MACROLET instead, as this does not cause the generation of
a debug function in the compiled code.