Hash-cons the specialized array element type properties data list
This resolves at least one portability problem: WIDETAG->ELEMENT-TYPE
generates a table with entries being the literal SAETP-SPECIFIER
values.  If those specifiers contain EQUAL-but-not-EQL data -- such as
for example the cons (32 . NIL) at the end of (SIGNED-BYTE 32)
and (UNSIGNED-BYTE 32) -- then compilers may choose whether or not to
coalesce those data to EQL structures when dumping during the
compilation of the xc host.  By hash-consing, we ensure that all EQUAL
substructures in the data are also EQL, removing this choice from the
host compilers.