Fix another cross-compiler portability issue
Well.  The real issue is that the return value of SUBSTITUTE-IF is
allowed to be EQL to, or not EQL to, the sequence it operates on if
the condition for substitution is never met.

The guard against calling SUBSTITUTE-IF was possibly in part meant to
deal with that, and only actually call substitute if there was
guaranteed to be work to be done.  (I'm not sure how strong that
guarantee is).  Unfortunately, the EQ check was done against the wrong
constant, meaning that the call to SUBSTITUTE-IF was effectively
always done, leading to the generation of a number of 62-element
vectors containing all-Ts.