Accept -jN arg to build-all-cores
and don't redirect *stdout* and *stderr* if -j1.

You can specify more than one target, and they will follow one another
with no chance of overlap since it's sequential. This should give more-or-less
the same behavior as the deleted crossbuild-runner/build-all.sh script.

We can add a few more indicators of where each build starts,
but this is probably adequate for now.