Put huffman into make-target-2 and simplify reading data
The logic in target-char which created binary-searchable vectors is
not needed after the most recent batch of changes in target-unicode.

Since the perfect hash generator receives its inputs in a different
order - though the keys are the same - this change wipes out and
rebuilds the cache file to purge old entries.