debug-info: Simplify &more handling.
c2404a2f430ecf57897a795202625dff4764c18d added the ability for the
debugger to display and bind more args to the sb-debug::more
symbol. However, we can do this by grabbing the more args debug vars
out of the debug fun directly instead of allocating 2 flag bits to
annotate every debug var. This allows the tag bits to be used for
other things and also removes the need for an extra info slot in
parsed debug var structures.