Compute values of parsed cross-floats
Rather than using the host's parser in a moderately-controlled way,
generate the exact rational corresponding to a specific decimal float
notation, and then compute the bits for the nearest IEEE 754 float of
the appropriate format.

This generates no new entries in the xfloat journal in and of itself.
We add the results of parsing to the xfloat journal in preparation for
a new implementation where we perform and journal portable float
computations on all hosts, and verify those computations in the built
sbcl.  For parsing floats, this is slightly wonky because the natural
key to cache the results under involves READ-FROM-STRING, but
READ-FROM-STRING is sensitive to *READ-DEFAULT-FLOAT-FORMAT* (and we
do bind *READ-DEFAULT-FLOAT-FORMAT* during the build in an attempt to
write code that is agnostic to the maximum target precision, because
one day we will support distinct LONG-FLOAT...)