Implement MIN/MAX on flonums portably
We need to make sure to match the host's interpretation of ordering in
case there are distinguishable but equal arguments, so (MIN -0.0 0.0)
and (MIN 0.0 -0.0) are consistent with the target.  (In practice this
case doesn't arise in cross-compilation, and additionally we have the
safety net of the check in warm load that all cross-float operations
are consistent.)

Now that MIN/MAX (and the underlying comparators) support comparisons
with infinities, remove the #+nil guarding an example that we could
not previously cross-compile.