Don't check emulated float computations against the host.
It's more relevant to check that the cross-float emulation produces
the same results that the actual target produces, which we make sure
to do at the end of the build.

Whether the host produces the same result or not is inconsequential to
the build artifacts. Indeed, older SBCL hosts do sometimes have
different float behavior and will always complain, and a host with a
differing float precision (e.g. if #+long-float is enabled somewhere)
will never pass the checks. This is unactionable noise: if an
architecture actually does produce incorrect results which mismatch
emulation, the post-build float verification step will complain with
an error.