Share a good amount of target float logic with cross-float.
* By doing some minor rearrangements of the build order, we can define
some of our target system float machinery used for float decoding and
infinity/nan handling for the host as well, and we have access to
float related constants and accessors, so we can share even more logic
in the future. This helps us with eventually removing the journal,
since its main function now is checking that cross-float and
target-float logic do the same thing.

* This gives us INTEGER-DECODE-FLOAT (and the rest of the float
decoding functions) automatically on the host. Use this in the
cross-compiler, so we can get more float-tran optimizations at
cross-compile time while ditching a very specific oracle. Also use it
to reimplement SB-XC:RATIONAL (so we can ditch custom flonum field
extractors).

* The float reader parts of cross-float are now distributed across
cross-early and a new file, to help with defining things in the right
order so we can do this.

* Add a docstring for FLOAT-DIGITS.