Add IGNORABLE declarations
Compiling src/code/thread.lisp using sbcl-1.2.7 as the host compiler
for single-threaded builds triggers a style warning, which breaks the
build at the end of make-host-2.  The compile-time effect of DEFMACRO,
even during cross-compilation, must generate a host function in order
to perform macroexpansion during cross-compile time, so we are
sensitive to the host's signals at that point, and apparently older
sbcls emitted style-warnings for unignored variables that were only
used in dead code.