Stop mutating CL:*READ-DEFAULT-FLOAT-FORMAT* during the build.
Sicne we now control how floats are read when building the system,
there is no need to frob the host's value
of *READ-DEFAULT-FLOAT-FORMAT*. We also represent target floats
portably now, so we can just spell floats how we intend to read
them (long floats aren't supported yet but that's just a matter of
adding them to cross-float; it's no longer impossible to read the
literals in portably like it was before we added a portable emulation
layer.)

Also add an SB-XC: shadowed version of the variablefor if we ever do
want to mess with how the (xc) reader reads floats by default. This
way we won't mutate the host's value.