Have a single closure for all hash table accessors
That is, one anonymous closure for each of GETHASH, PUTHASH, and
REMHASH. This has no discernible effect on microbenchmarks while
making the code smaller and marginally speeding up things. This is in
preparation for adaptive hashing, which calls PICK-TABLE-METHODS more
often.

The downside of this change is that the closures are not longer named
(i.e. GETHASH/EQUAL/LOCK), which is not a big deal because the next
frame on the stack is still named (i.e. GETHASH/EQUAL).