Waste less time and memory filling in LAYOUT-SLOT-MAPPER
Build the map of slot-name to DSD-INDEX only when SLOT-VALUE is called
on a structure instance, which delays MAKE-PERFECT-HASH-LAMBDA until needed.
The caller of SLOT-VALUE is not blocked waiting for findhash or COMPILE
because a trivial (slower) version of the slot mapper is always installed
right away. The slower way is simply linear scan of slot names.
An optimized function is computed and installed by the finalizer thread.

This reduced by about 30% the time to load (as source) hundreds of
lisp files defining thousands of structure types.