Stop passing an alien-linkage-table address to pthread_create
Alien linkage entries are for Lisp-to-C call, and not for C-to-C call.
If you change arch_write_linkage_table_entry to produce instructions
that assume anything about Lisp state, such as that the thread-base
register points to the current primitive thread (say for per-thread
call-out profiling), it could crash in the new thread because in all
likelihood the assumed thread-base register contains a wrong value.