Improve package iteration per lp#2080387
Allow changing accessibility from :INTERNAL to :EXTERNAL during iteration
without messing up the iterator. But now we have a duplicates problem which
it's unclear whether runs slightly afoul of the spec. e.g. DO-ALL-SYMBOLS
"may cause a symbol that is present in several packages to be processed more
than once". Since that's not phrased as if-and-only-if, we take the liberty
of returning a symbol more than once even when NOT present in more than
one package. Apparently so do some other Common-Lisp implementations.