Adaptive hashing: EQ hash tables
When an EQ hash table reaches a certain size, it switches to a very
simple hash function that has the following properties:

- It is a perfect hash for strided ranges of fixnums (e.g. 0*k, 1*k,
  2*k 3*k, 4*k, ...), where k is an integer.

- For non-immediate objects, it is somewhat cache-friendly because it
  hashes nearby addresses more densely into buckets.

This hash function is actually

    (+ (ash address n)
       (ash address -16))

where N is a parameter determined heuristically by calculating the
number of trailing bits common to all key addresses at the time of the
switch. 16 is about the number of bits in a gencgc page plus 1.