Make a way to compile 1 sexpr into a fasl
It could kinda have been done already, but if your form contained
dumpable-though-unprintable subparts, it wasn't fun to produce the
source file for that. (It needs a reader macro to return the form
as-is such as using "#." which works but is ugly)
So this adds a convenient interface, not to mention makes it easier
to get an IR trace than by binding *COMPILER-TRACE-OUTPUT* by hand.