Propagate "undefined function" warnings to load-time
This is intended to supplant the concept of deferred-warnings in ASDF by
imparting the ability to the FASLoader to issue warnings about unresolved
function linkage. At last, Lisp can use a more modern paradigm of
separately compiled translation units.

LOAD is totally unchanged if you don't hook into this mechanism in user code.