Avoid leaking zombie processes on Unix due to PROCESS-CLOSE.
PROCESS-CLOSE has always removed the process from *ACTIVE-PROCESSES*,
but prior to 1e7fc473, the SIGCHLD handler's use of wait3() prevented
zombies. The switch from wait3() to waitpid() in 1e7fc473 subtly
changed the meaning of *ACTIVE-PROCESSES* from "the processes SBCL
updates" to "the processes SBCL waits for and updates": after
1e7fc473, nothing has waited for a process that's been
PROCESS-CLOSE'd. So this commit has PROCESS-CLOSE leave the process
in *ACTIVE-PROCESSES* and toggle a separate slot to T.

On Windows, I believe closing the process handle is the way to avoid
zombies, so PROCESS-CLOSE can remove a process there.

committer: Stas Boukarev <stassats@gmail.com>