Fix broken class tests for [FUNCALLABLE-]STANDARD-OBJECT
In discriminating nets, the class tests for STANDARD-OBJECT and
FUNCALLABLE-STANDARD-OBJECT used the "low"-level STD-INSTANCE-P and
FSC-INSTANCE-P, which despite their names are not suitable for general
use but instead admit false positive (they test for INSTANCE and
FUNCALLABLE-INSTANCE respectively).  This meant that they would
incorrectly classify e.g. structures, conditions, and arbitrary
funcallable-instances as STANDARD-OBJECT.

Mostly the discriminating net code is latent.  It is possible to
convince it to come into play, for example in generic functions with
very large numbers of required arguments, or with EQL specializers
or (in some other cases) as part of a secondary dispatch function.
Add some direct tests of the discriminating net code as well as
implicit tests through pure "user code".

Also remove some other uses of STD-INSTANCE-P and FSC-INSTANCE-P where
the context makes it clear that their system-low-level equivalents
should be used directly.