arm64: More robustly implement after-breakpoints.
We need to interpret the instruction replaced by a breakpoint in order
to figure out where to place the after-breakpoint, so that we can
reinstall the breakpoint. The instruction may be a control flow
instruction so compute the target instruction. Very much untested
because breakpoints aren't usually inserted at control flow
instructions.