Finish #+sw-int-avoidance patch less ambitiously
It supports only pending stop-for-GC now, but if debugging a genuine SIGILL
or SIGTRAP, it's a lot easier when you can avoid most GC-related signals.
(It's easy enough to tell gdb to ignore/pass SIGUSR2, but not SIGILL or
SIGTRAP if you're expecting to receive and stop at those)
So this is a different take on rev aa44aa035f but better, in that the
entire regression suite on Linux succeeds.
The macOS implementation is just a sketch; it doesn't work at all.