Revert "stack: Use lexical scoping to extend stack object lifetimes."
There exists a large body of code for which this change causes
problems.

It's very possible that the problem lies in the large body of code
rather than in this specific change, but other factors have prevented
establishing this one way or the other.  Reverting this for this
release cycle, with the plan to reintroduce it early in the next
cycle and work on the other factors, is I think the least bad option.

This reverts commit 66ee531ea7a3853da832c57ebf68b447162b8029.