x86-64: tweak #+mark-region-gc for presumptive static space relocation
The main problem presented by #+mark-region-gc (or #-immobile-space) is use of
of hardwired addresses for linkage tables, both lisp and alien calling.
This patch remedies that by referencing linkage entries off NULL-TN.

Some immobile-space hacks are removed from the machine-independent pieces
of coreparse and instead done in a *-arch.c file which has a new function
named coreparse_alloc_space() that is intended to be customized as needed.
The basic implementation remains in coreparse for other architectures.

This change still does not actually allow relocating static space, and in fact
takes a step backwards by disabling the elf-sans-immobile test until I finish
relocatability. However in triaging the current state, I discovered a more
urgent concern: as static space is of variable size based on dynamic-space and
hence the GC card table size, and not made by allocate_hardwired_spaces(),
is it _less_ likely that the OS will place it as requested.
So that becomes the most important issue to resolve at the moment.