Fix DEBUG_CORE_LOADING more
The purpose of the load-time "sanity" pass being to verify that the
final GC in save-lisp-and-die did not preserve unreachable objects,
all weak objects should be skipped over. Doing that exposed a
deficiency in codeblob tracing, and all the extra printing about
unreachable functions made it hard to find any unreachable SAPs
which is what I was looking for in the first place.
(None of this matters to users- the #define is for developers.)