Fix adjacent gc barriers elimination.
impart_mark_stickiness skipped pointers into gen 0, which is fine if
both the value and the object are allocated before the card mark is
stored. Eliminating a barrier may insert an allocation sequence
between two stores, which may cause the value and the object to have
different generations.
The extra cost of looking at more card marks should not be significant
and reducing code size and performance of consequent stores is
valuable.