Remove :alien-{code,data}-linkage-index fixup flavors
:foreign[-dataref] are enough, though we need different techniques to apply the
fixup depending on immobile code or not. This "should" be captured in the KIND
of fixup, but it's disappointingly tricky to adapt the assembler to different
kinds with the same flavor. I'm also annoyed that this adds more logic to the
disassembler, however, the change resolves a problem cited in the comment above
FOREIGN-SYMBOL-SAP - the linkage stub should entail a yieldpoint-based GC stop,
which would not work if some uses of the entry rely on the abstraction leakage
that "CALL [x+8]" is nearly the same as "CALL x".