Fix major misunderstanding about :ABS32 fixups
I should never have made NIL-relative fixups "exceptionally" take a signed
value. On the contrary, the weird case is a :DWORD operand and a 32-bit
unsigned value because in any operation on a 64-bit value, if the CPU wants
a 32-bit immediate operand, the immediate is signed-extended.

This was technically wrong from the moment the 64-bit port first came to life,
but calling to static space could treat (SIGNED-BYTE 32) and (UNSIGNED-BYTE 31)
as basically the same due to the sub-2GB address constraint.

Well now the logic says what it means and means what it says.