Recursively look for stack allocatable parts in one pass.
Instead of doing multiple passes to check if we can actually stack
allocate before and while committing to setting the dynamic extents of
all affected lvars, merge the logic of setting the dynamic extents
with the logic for checking whether it's possible to stack
allocate. This makes the logic simpler (at the very least,
concentrated in one place) and avoids extra work recursing through the
same data flow over and over again.