Give a choice of assigning all DEFVARs a load-time TLS index
For x86-64 this reduces emit-symeval to 3 instructions worst case, using fewer
abs32 fixups. It also cures a reproducibility problem stemming from the fact
that TLS index can be wired in for "... any symbol the compiler decides ... due
to compiling".  Specifically, if file 1 contains a DEFVAR, file 2 has a use of
it, and file 3 has a LET binding, the artifacts for files 2 and 3 should not
be changed if flipped in compilation order, assuming no interdependence.