Auto-stack-allocate named local funs whose single refs don't escape.
A LEXENV-CONTAINS-LAMBDA test was originally added which prevented
most cases of this from working in order to prevent stack allocated
closures from getting cleaned up too many times, which also made it
necessary to try to dxify downward funargs before any transformation
happens. However, stack analysis has improved significantly since
then (specifically because we now do a pass to insert
%DYNAMIC-EXTENT-START explicitly), such that The Right Thing is just
done for us by stack analysis in terms of making the extent come out
right in the presence of stuff like loops in the CFG. We only have to
make sure that the closure allocation actually happens in the same
stack frame as the reference (where the cleanup will get
inserted). See the tests around AUTO-DX-CLEANED-UP-TOO-MANY-TIMES,
which can now be asserted to not cons as well. Since this is a common
pattern, we can expect this to help in real code.

In order to accomplish this, we need to change the representation of
how ENCLOSE and DYNAMIC-EXTENT work, now that multiple dynamic extents
can be derived for the same ENCLOSE. We essentially union all derived
local function dynamic extents for stack analysis.

The functionals stack allocated by the ENCLOSE node should only get
cleaned up when all associated derived dynamic extents end. We
represent this in the flow graph by making a distinction between
derived dynamic extents and declared dynamic extents, preferring the
declared dynamic extent when it is available. This fixes some issues
with auto-stack-allocation where cleanup code was inserted too early
due to two functions sharing the same cleanup.

This also paves the way to allow multiple reference local functions to
get auto-stack-allocated as well, for which a failing test has been
added.