arm64: Make breakpoints thread-safe.
We achieve this by using a different approach for
DO-DISPLACED-INST. Instead of using after-breakpoint-traps, we just
simulate pc-relative instructions directly or otherwise copy the
displaced instruction into an on-demand allocated thread-local
trampoline that jumps back to the right place. This also removes the
overhead of an additional trap.

This combined with 484ffd9bde2a38ae56495f8065649f06f9b44946 makes
breakpoint tracing thread-safe as well, fixing lp#1837306 on
arm64. Other architectures which pin code on the stack could do the
same provided that code to simulate pc-relative instructions are
written.