Bootstrap pcl w.r.t. method tracing differently.
Instead of messing around with the build order to let PCL hook into
trace because of bootstrap reasons, just define MAYBE-TRACE-METHOD
properly once the rest of the tracing facilities are defined. To
bootstrap PCL correctly, we can install a dummy bootstrap definition
that gets redefined later. This is nicer than checking whether a hash
table is empty as a proxy/workaround. Unfortunately, for reasons I
don't understand, we still need to do that for unrelated reasons
involving failed mop tests, so that is probably unexpected behavior
that should be fixed. We can get rid of early-ntrace.lisp now though.