stack: Delete some unreachable code.
There's no need to handle exit-lvar's specially. They are annotated in
PUSHED sets just like lvar's of any other combinations, since the
%NLX-ENTRY stub has been inserted. This also removes an assumption on
the exact representation of the %NLX-ENTRY stub (namely that it is the
last node in its block).

Also don't guard a cheap AVER with *check-consistency*. Similar AVERs
in stack analysis are always checked.