envanal: Fix otherwise inaccessible subparts logic + cleanup.
By recursing and using the lvar's dynamic extent slot to return
whether there was a deeper subpart, it could be the case that for
combinations which deliver their values to the same lvar, stack
allocating the result of one combination could impact the stack
allocation decision for the arguments of the other combination even
when it isn't warranted. Therefore we have to recursively return and
check when a deeper subpart was actually found for the recursion.

This also makes the logic for propagating dynamic extents in the
presence of another dynamic extent more obviously correct.

Also remove two impossible-to-satisfy conditions:
dynamic-extent-cleanups are never null after initialization now, and
the info slot should also be uninitialized after looking for stack
allocatable parts per dynamic extent.