Implement the new frlock (or "seqlock") strategy
For now I've just conditionally compiled a new implementation of the macros
so that the regression test runs unchanged. Wikipedia says that in Linux the name
was frlock but now is seqlock, so maybe we should do the same and name the
interface SEQLOCK-READ, SEQLOCK-WRITE so that we can offer both old and new
in case people "need" the old one. Especially as we documented functions which
access counter fields that do not exist in the new logic.