Allow main thread to exit despite attachment of remote ldb peer
An fgets in ldb_service_main can block _IO_cleanup (in libc) from finishing
in the main thread, which is a slightly surprising reason for an SBCL process
to appear hung. The main thread would need to kill the ldb thread to fix that,
so instead just use the variant of fgets that does not acquire a stdio lock.
This affects nobody, since no user should define START_LDB_SERVICE_THREAD.