Create main SB-THREAD:THREAD before starting lisp initial function
i.e. make a 'struct thread_instance' in C. If we can arrange for the two required
physical structures to coexist for all live threads, then deciding to add a
field to or the other becomes a decision only based on where it is most used,
and not when its containing structure first comes into being. A kludge in
thread_accrue_stw_time remains temporarily, because of this exact problem.