Streamline UNIX-WRITE a litte
(ADDR (DEREF ...)) is a very roundabout way of computing the same result
that SAP+ does, except the former allocate a local alien at runtime.
Despite the new code always binding *PINNED-OBJECTS* on the precise backends,
it's smaller and quicker, especially as 20x more calls to UNIX-WRITE get
an array instead of a SAP in typical use, meaning a pin had to occur.
(Profiling shows array of base-char and UB8 as the commonest cases)