Fix SB-COVER when reading a read-conditionalized top-level form
When reading sources to generate a coverage report, if a form is
preceded by a positive read-time feature conditional, both the
conditional and the form itself will have entries for the object
corresponding to the form.  It's therefore completely normal for there
to be more than one position associated with that form, and we want
the innermost (last) one.

(This problem also showed up for internal forms, but the coverage
would then simply move on to the next outer one: that's not possible
for a top-level form.)