Add some tests
This matches current behaviour, which is neither what I expected nor
what was documented in the original (lp#1361502) bug report, but at
least helps to see what will change as and when we change the
implementation of source paths for commas.  If nothing else, the
triple-comma test demonstrates a lack of generality in the current
implementation.