Remove some recursive IR1-attributes
It might be nice to shrink the number of functions with recursive
attributes even more, particularly in the numeric domain by (perhaps)
having specialized functions for the typical one-level-of-recursion
decomposition of transcendental functions on complexes in terms of
simpler operations on reals.  However, this removal of recursive
attributes would have caught the ATANH miscompilation when compiling
the system with coverage marks.