Rewrite source-path generation for commas
When generating source-paths, treat a comma struct as representing a
list (comma <expr>), handling the cases when the comma is the bare
CDR or recursion down the CAR of a list.  Adapt getting the original
source for a source path to handle this convention, and
form-number-translations likewise.

This allows sb-cover to define custom comma and backq reader macros
which actually generate this list structure, making the simple
list-walking to find character positions there correct.  (It also
allows a minor modification to SLIME along the same lines to restore
buffer annotations at the right place in backquoted forms.)

Include tests of SB-COVER on backquoted / unquoted forms, and an
"incompatible change" note in NEWS.