Don't special-case QUOTE in source path finding
There is some danger that removing this will cause slowdowns in
compiling code with large list/tree literal constants -- see lp#654289
for one example of this.  It's probably the case that embedding large
list/tree literal constants is A Bad Idea for other reasons -- for
many of the same reasons as with structured objects in DEFCONSTANT, so
we should probably enhance the source tracking for forms involving
quotes inside backquotes rather than speed of compiling code with
somewhat dubious list constants.