Omit some do-nothing string-to-c-string calls
The runtime does not cons a fresh simple-base-string when passing one to
a C function declared to receive UTF8-STRING, however the compiler didn't
realize that this situation does not need a runtime conversion.
No regression test exercised a declared C argument type of UTF8-STRING
(as defined in target-c-call) so this adds one.