Add bunch of commentary about small READ bug re char traits
I doubted the FIXME that said "I doubt they're in the manual" so I hunted
down a relevant CLHS section explaining character traits, whereupon I saw
that #\| is intended to be slightly less powerful than #\\ because #\\
allows "invalid" constituents into a token but #\| should not. (news to me)

And delete mention of READ-UNQUALIFIED-TOKEN which was deleted or renamed
before branching from CMUCL because it's not in there either.