x86-64: Disable use of ICEBP instruction for breakpoints
It sometimes but not always makes the sigcontext contain the wrong
saved RIP under virtualization (e.g. cloud compute platforms)

If using INTO instructions to signal SIGILL for error breaks, then use UD2
for receive-pending-interrupt. This makes program-analyzing programs think
our asm code is well-formed and not full of random bytes in most places.

For the record, the current state, and history, is as follows:
- By default we use 0xCC traps. This was apparently a problem for #+darwin,
  at least back in 2010. So git rev 11b5ac86 (or before) added UD2 as a choice
- Rev 9cca9773 allowed 0xCE as a choice, trigged by the INTO instruction to
  deliver SIGILL. (SIGTRAP annoys gdb which wants SIGTRAP for itself. And
  note that INTO delivers SIGTRAP in 32-bit code, but SIGILL in 64-bit
  because the instruction was made invalid)
- Rev 2fb2d485 removed UD2 as the default under macOS, as it's not needed.
  And I don't believe UD2 was ever the default for anything except macOS.

src/cold/shared prevents trying to enable both :int4- and :ud2-breakpoints.
The only thing that precludes making 0xCE the default is fear- technically
it is not _defined_ to deliver SIGILL, but is "merely" illegal (and might
be reasigned?), whereas UD2 is literally permanently reserved as illegal.