x86-64: Improve break instruction change a little more
Always use unadorned UD2 for pseudo-atomic unless we can't.
The only reason to avoid it for receive-pending-interrupt is that UD2 might
be needed for _all_ traps, which in general are followed by opaque bytes.
And officially there is no single-byte illegal opcode, which we'd like
to have because reasons. (So claim revs 795e122b and 6e722b9c regarding
Haiku OS though the situation may have changed since.) But anyway, if UD2
isn't used for other purposes, then using an instruction that is not
followed by un-decodable bytes is preferable to one that is.

Back story: we got a complaint that our Lisp binaries contained a lot of
AVX instructions that were causing CPU throttling, which may affect tasks
other than your own, due to thermal management. So I promised we don't have
a ton of AVX instructions, and that any offline tool which statically analyses
Lisp assembly code is going to wrongly decode due to loss of sync after every
0xCC byte unless you're privy to the magic of the Lisp payload bytes.