Process smashed weak hash-table cells more quickly
The first (SETF GETHASH) operation on a given table after a GC now deletes
the entire list that GC constructed to transfer the indices of culled cells
to Lisp. This is not great because the list is consed in alloc_generation
following any particular collect_generation. That is, we can't always
allocate to generation 0. Nor can the construction of such list be delayed
until just before returning from collect_garbage, because that would lose
information about what was clobbered. So at worst, we still cons the list
in a high generation only because we are forced to.

There are a couple of solutions but none readily at hand.
My sense is that we should change weak tables to using an open-addressing
strategy wherein it would be legal to stuff tombstones into k/v cells
during GC, and have probing figure out what happened. (Like maybe
rehashing the table if there are too many tombstones)